requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
idna==3.7
platformdirs==4.2.0
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
